Fix MALLOCX_ALIGN type warnings

+++ modules/core/src/main/include/system/jemalloc/jemalloc.h
@@ -92,7 +92,7 @@ extern "C" {
 #    define MALLOCX_ALIGN(a)	(ffs(a)-1)
 #  else
 #    define MALLOCX_ALIGN(a)						\
-	 ((a < (size_t)INT_MAX) ? ffs(a)-1 : ffs((size_t)a>>32)+31)
+	 ((a < (size_t)INT_MAX) ? ffs((unsigned int)a)-1 : ffs(a>>32)+31)
 #  endif
 #  define MALLOCX_ZERO	((int)0x40)
 /*
